[FLINK-10041] Extract iterators from RocksDBKeyedStateBackend (inner or static inner classes) into full classes
